requirements.txt
DateTime~=5.1
colorama~=0.4.6
Werkzeug~=2.3.7
numpy~=1.24.3
scipy~=1.10.1
pydub~=0.25.1
Flask~=2.3.3
flask_session~=0.5.0
Jinja2~=3.1.2
google-cloud-storage~=2.10.0
boto3~=1.34.3
paramiko~=3.4.0
tweepy~=4.14.0
gunicorn~=21.2.0
pika~=1.3.2
redis~=5.0.0
mysql-connector-python~=8.0.33
requests~=2.31.0
bcrypt~=4.1.2